requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.7.0
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
adwaita-icon-theme=49.0=unix_0
alembic=1.17.0=py311h06a4308_0
asttokens=3.0.0=pyhd8ed1ab_1
at-spi2-atk=2.38.0=h0630a04_3
at-spi2-core=2.40.3=h0630a04_0
atk-1.0=2.38.0=h04ea711_2
blas=1.0=openblas
brotli=1.2.0=h41a2e66_0
brotli-bin=1.2.0=hf2c8021_0
bzip2=1.0.8=hda65f42_8
ca-certificates=2025.10.5=hbd8a1cb_0
cairo=1.18.4=h3394656_0
charset-normalizer=3.4.4=pyhd8ed1ab_0
colorlog=5.0.1=py311h06a4308_1
comm=0.2.3=pyhe01879c_0
contourpy=1.3.3=py311hdf67eae_2
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.28=hd9c7081_0
dbus=1.16.2=h3c4dab8_0
debugpy=1.8.17=py311hc665b79_0
decorator=5.2.1=pyhd8ed1ab_0
epoxy=1.5.10=hb03c661_2
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.60.1=py311h3778330_0
freetype=2.14.1=ha770c72_0
freetype-py=2.3.0=pyhd8ed1ab_0
fribidi=1.0.16=hb03c661_0
gdk-pixbuf=2.44.4=h2b0a6b4_0
glib-tools=2.86.0=hf516916_1
graphite2=1.3.14=hecca717_2
graphviz=13.1.2=h87b6fe6_0
greenlet=3.2.4=py311h1ddb823_1
gtk3=3.24.43=h993cebd_6
gts=0.7.6=h977cf35_4
harfbuzz=12.1.0=h15599e2_0
hicolor-icon-theme=0.17=ha770c72_2
icu=75.1=he02047a_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=7.1.0=pyha191276_0
ipython=9.6.0=pyhfa0c392_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
joblib=1.5.2=py311h06a4308_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.9.1=pyhc90fa1f_0
keyutils=1.6.3=hb9d3cd8_0
kiwisolver=1.4.9=py311h724c32c_1
krb5=1.21.3=h659f571_0
lcms2=2.17=h717163a_0
ld_impl_linux-64=2.44=h1aa0949_4
lerc=4.0.0=h0aef613_1
libboost=1.86.0=hed09d94_4
libboost-python=1.86.0=py311h1d5f577_4
libbrotlicommon=1.2.0=h09219d5_0
libbrotlidec=1.2.0=hd53d788_0
libbrotlienc=1.2.0=h02bd7ab_0
libcups=2.3.3=hb8b1518_5
libdeflate=1.24=h86f0d12_0
libdrm=2.4.125=hb03c661_1
libedit=3.1.20250104=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libegl-devel=1.7.0=ha4b6fd6_2
libexpat=2.7.1=hecca717_0
libffi=3.5.2=h9ec8514_0
libfreetype=2.14.1=ha770c72_0
libfreetype6=2.14.1=h73754d4_0
libgcc=15.2.0=h767d61c_7
libgcc-ng=15.2.0=h69a702a_7
libgd=2.3.3=h6f5c62b_11
libgfortran-ng=8.2.0=hdf63c60_1
libgfortran5=15.2.0=hcd61629_7
libgl=1.7.0=ha4b6fd6_2
libgl-devel=1.7.0=ha4b6fd6_2
libglib=2.86.0=h32235b2_1
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libglx-devel=1.7.0=ha4b6fd6_2
libgomp=15.2.0=h767d61c_7
libiconv=1.18=h3b78370_2
libjpeg-turbo=3.1.0=hb9d3cd8_0
liblzma=5.8.1=hb9d3cd8_2
libnsl=2.0.1=hb9d3cd8_1
libntlm=1.8=hb9d3cd8_0
libopenblas=0.3.30=h46f56fc_0
libpciaccess=0.18=hb9d3cd8_0
libpng=1.6.50=h421ea60_1
libpq=18.0=h3675c94_0
librdkit=2025.09.1=h3c5c181_1
librsvg=2.60.0=h61e6d4b_0
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.50.4=h0c1763c_0
libstdcxx=15.2.0=h8f9b012_7
libstdcxx-ng=15.2.0=h4852527_7
libtiff=4.7.1=h8261f1e_0
libuuid=2.41.2=he9a06e4_0
libwebp-base=1.6.0=hd42ef1d_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.12.2=hca5e8e5_0
libxml2=2.15.1=h26afc86_0
libxml2-16=2.15.1=ha9997c6_0
libzlib=1.3.1=hb25bd0a_0
llvmlite=0.44.0=py311hc1e8f15_1
mako=1.3.10=py311h06a4308_0
markupsafe=3.0.2=py311h5eee18b_0
matplotlib-base=3.10.7=py311h0f3be63_0
matplotlib-inline=0.2.1=pyhd8ed1ab_0
munkres=1.1.4=pyhd8ed1ab_1
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.5=py311h06a4308_0
numba=0.61.2=py311h6a678d5_0
numpy=2.2.5=py311hc213724_2
numpy-base=2.2.5=py311hf387b04_2
openjpeg=2.5.4=h55fea9a_0
openldap=2.6.10=he970967_0
openssl=3.5.4=h26f9b46_0
optuna=4.5.0=pyhd8ed1ab_0
packaging=25.0=pyh29332c3_1
pandas=2.3.3=py311hed34c8f_1
pango=1.56.4=hadf4263_0
parso=0.8.5=pyhcf101f3_0
pcre2=10.46=h1321c63_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=py311h98278a2_3
pip=25.2=pyh8b19718_0
pixman=0.46.4=h54a6638_1
platformdirs=4.5.0=pyhcf101f3_0
prompt-toolkit=3.0.52=pyha770c72_0
psutil=7.1.1=py311h49ec1c0_0
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycairo=1.28.0=py311hc1a9592_1
pygments=2.19.2=pyhd8ed1ab_0
pygraphviz=1.14=py311hf824169_2
pyparsing=3.2.5=pyhcf101f3_0
python=3.11.14=hd63d673_2_cpython
python-dateutil=2.9.0.post0=pyhe01879c_2
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.11=8_cp311
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0.2=py311h5eee18b_0
pyzmq=27.1.0=py311h2315fbb_0
qhull=2020.2=h434a139_5
rdkit=2025.09.1=py311h34e1bb8_1
readline=8.2=h8c095d6_2
reportlab=4.4.4=py311h49ec1c0_0
rlpycairo=0.4.0=pyh6c17108_0
scikit-learn=1.7.2=py311h505adc9_0
scipy=1.16.2=py311ha592af2_0
setuptools=80.9.0=pyhff2d567_0
six=1.17.0=pyhe01879c_1
sqlalchemy=2.0.44=py311h49ec1c0_0
stack_data=0.6.3=pyhd8ed1ab_1
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py311h92b7b1e_0
tk=8.6.13=noxft_hd72426e_102
tornado=6.5.2=py311h49ec1c0_1
tqdm=4.67.1=py311h92b7b1e_0
traitlets=5.14.3=pyhd8ed1ab_1
typing-extensions=4.15.0=h396c80c_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py311h49ec1c0_1
wayland=1.24.0=hd6090a7_1
wcwidth=0.2.14=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
xkeyboard-config=2.46=hb03c661_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.6=he73a12e_0
xorg-libx11=1.8.12=h4f16b4b_0
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.2=hb03c661_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxinerama=1.1.5=h5888daf_1
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
xorg-xorgproto=2024.1=hb9d3cd8_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h387f397_9
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.3.1=hb25bd0a_0
zstd=1.5.7=hb8e6e7a_2
